move ObjectExtensionTests to seperate file
